Show generic recursive navigation proof-of-concept
Implements a pure-Liquid recursive navigation tree, as described in
navlist.html. This version is meant to be generic, so for instance it
always renders all root pages (siblings of /). The actual
requirements for our City of Austin Pages use case are more specific
(see subsequent commits for details), but I figured I'd commit the
initial proof-of-concept for posterity.