Customize recursive navigation to support multiple micro-sites
Effectively an opinionated version of the generic recursive navigation
in the previous commit. The key difference is the introduction of a
"toc_level" (Table of Contents level), a front-matter property that
configures the URL-level at which "micro-sites" (i.e. individual
Guides and Projects) exist. Each micro-site gets its own header and
persistent, fully expanded navigation.