Create an AWS Lambda function to trigger gh-pages updates
The previous deploy.sh script worked fine, but setting up a GitHub
webhook to trigger CircleCI builds directly from submodule
repos presented a problem, since it's not possible to only trigger a
webhook/build on changes to the submodule master branch (GitHub's push
event fires on every push).

This solution implements a webhook receiver "API" that runs on AWS
Lambda and API Gateway, and filters and translates the webhooks before
effectively proxying the request to CircleCI. As a consequence, the
deploy.sh script that runs on Circle is now much "smarter": it updates
only the affected submodule, rather than all of them, and it includes
a GitHub-compatible link to the commit in the source repo right in the
gh-pages commit.

This was my first time doing much of anything with Lambda, so there
were a lot of lessons learned, which I might cover in more depth in a
blog post or more general documentations. For now, some highlights:

* Lambda doesn't support env variables, which rules out the typical
  12-factor secret storage. The solution: use Amazon's KMS service to
  encrypt the secrets, which can then be safely stored directly in
  the repo (even if it's public). I used an npm package called
  [Cryptex][] to help.

* There is a wide variety of dev/deploy tools for Lambda functions. I
  looked at Claudia, Gordon, and Serverless, and used Claudia for
  initial deployment and testing, but found those options lacking in
  one way or another. I went with the decidedly more-verbose, less-
  magical, and wonderfully predictable [Terraform][], which is a more
  general purpose tool and might seem like overkill, but I think it's
  great. The Lambda function that Terraform uploads is created by
  [node-lambda][], via `npm run pack`.

[Cryptex]: http://technologyadvice.github.io/lock-up-your-customer-accounts-give-away-the-key/
[Terraform]: https://www.terraform.io
[node-lambda]: https://github.com/motdotla/node-lambda

TODO:

* Add guidance about working with api/ to a README

Closes #10